patch/configs-microsoft-graph

+++ reviewer/src/main/java/com/reviewer/reviewer/controllers/FormController.java
@@ -1,28 +1,21 @@
 package com.reviewer.reviewer.controllers;
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import com.reviewer.reviewer.dto.forms.FormUpdateDto;
 import com.reviewer.reviewer.dto.forms.QuestionFormCreateDto;
 import com.reviewer.reviewer.dto.forms.QuestionFormResponseDto;
+import com.reviewer.reviewer.infra.functionsConfig.CompareRoles;
 import com.reviewer.reviewer.services.FormService;
-
-
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @CrossOrigin
@@ -33,35 +26,48 @@ public class FormController {
     @Autowired
     private FormService service;
 
+    @Autowired
+    private CompareRoles roles;
+
     @PostMapping
     @Transactional
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity<QuestionFormResponseDto> createForm(@RequestBody @Valid QuestionFormCreateDto data){
+    public ResponseEntity<QuestionFormResponseDto> createForm(@RequestBody @Valid QuestionFormCreateDto data, @AuthenticationPrincipal Jwt tokenJWT){
+        if (!roles.compareRoles(tokenJWT).equals("ROLE_ADMIN")){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         return ResponseEntity.status(HttpStatus.CREATED).body(service.create(data));
 
     }
     @GetMapping("/{formId}")
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity<List<QuestionFormResponseDto>> listFormQuestion(@PathVariable(name = "formId") Long formId){
+    public ResponseEntity<List<QuestionFormResponseDto>> listFormQuestion(@PathVariable(name = "formId") Long formId, @AuthenticationPrincipal Jwt tokenJWT){
+        if (!roles.compareRoles(tokenJWT).equals("ROLE_ADMIN")){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         return ResponseEntity.status(HttpStatus.OK).body(service.listFormQuestion(formId));
 
     }
     @GetMapping
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity<List<QuestionFormResponseDto>> findAll(){
+    public ResponseEntity<List<QuestionFormResponseDto>> findAll(@AuthenticationPrincipal Jwt tokenJWT){
+        if (!roles.compareRoles(tokenJWT).equals("ROLE_ADMIN")){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         return ResponseEntity.status(HttpStatus.OK).body(service.findAll());
 
     }
     @DeleteMapping("/{formId}")
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity<?> deleteForm(@PathVariable(name = "formId") Long formId){
+    public ResponseEntity<?> deleteForm(@PathVariable(name = "formId") Long formId, @AuthenticationPrincipal Jwt tokenJWT){
+        if (!roles.compareRoles(tokenJWT).equals("ROLE_ADMIN")){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         service.deleteForm(formId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 
     }
     @PatchMapping("/{formId}")
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity<?> updateForm(@PathVariable(name = "formId") Long formId, @RequestBody @Valid FormUpdateDto data){
+    public ResponseEntity<?> updateForm(@PathVariable(name = "formId") Long formId, @RequestBody @Valid FormUpdateDto data, @AuthenticationPrincipal Jwt tokenJWT){
+        if (!roles.compareRoles(tokenJWT).equals("ROLE_ADMIN")){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         service.updateForm(formId, data);
         return ResponseEntity.status(HttpStatus.CREATED).body("Form edited successfully!");
